+++ frontend/src/components/RelativeTime.tsx
@@ -1,6 +1,16 @@
-import { parseDateBestEffortForRelative } from '../../utils/helpers';
+import React, { useEffect, useState } from 'react';
+import { formatAbsoluteTooltipAEST, parseDateBestEffortForRelative } from '../utils/helpers';
 
 export type RelativeTimeLocale = Intl.LocalesArgument;
+
+export type RelativeTimeProps = {
+  value: string | number | Date | null | undefined;
+  className?: string;
+  /** How often to refresh the label (e.g. so “5 minutes ago” advances). */
+  tickMs?: number;
+  /** Passed to `Intl.RelativeTimeFormat` (defaults to runtime locale). */
+  locale?: RelativeTimeLocale;
+};
 
 /** ISO string for HTML `dateTime` (UTC), using the same instant as relative text when strings are ambiguous. */
 export function toIsoTimestamp(
@@ -52,4 +62,32 @@ export function formatRelativeTime(
   }
   const diffYear = Math.trunc(diffMs / 31_536_000_000);
   return rtf.format(diffYear, 'year');
+}
+
+/** Default tick keeps second-level labels accurate (see formatRelativeTime). */
+const DEFAULT_TICK_MS = 1_000;
+
+export function RelativeTime({
+  value,
+  className,
+  tickMs = DEFAULT_TICK_MS,
+  locale,
+}: RelativeTimeProps) {
+  const [now, setNow] = useState(() => new Date());
+  useEffect(() => {
+    const id = window.setInterval(() => setNow(new Date()), tickMs);
+    return () => window.clearInterval(id);
+  }, [tickMs]);
+  const text = formatRelativeTime(value, now, locale);
+  if (text === '-') {
+    return <span className={className}>-</span>;
+  }
+  const nowMs = now.getTime();
+  const title = formatAbsoluteTooltipAEST(value);
+  const iso = toIsoTimestamp(value, nowMs);
+  return (
+    <time dateTime={iso} title={title || undefined} className={className}>
+      {text}
+    </time>
+  );
 }
