--- frontend/src/components/RelativeTime/index.ts
@@ -1,4 +1,0 @@
-export { RelativeTime } from './RelativeTime';
-export type { RelativeTimeProps } from './RelativeTime.types';
-export { formatRelativeTime, toIsoTimestamp } from './relativeTimeFormat';
-export type { RelativeTimeLocale } from './relativeTimeFormat';
